Delta (Delta v0.1.0)

Delta Elixir API

Configuration

Delta uses MongoDB for persistence and Mnesia for caching.

To configure Mnesia storage, add the following to your config:

config :mnesia, dir: './data/#{config_env()}/#{node()}'

Note that the :dir is charlist, not a string. If directory does not exist, it will be created.

To configure MongoDB connection params, set the :mongo_params key in your config:

config :delta, :mongo_params, [
  database: "delta_#{config_env()}",
  hostname: "localhost",
  port: 27017
]

More information on MongoDB connection params can be found in the mongodb-driver docs

Types

uuid4()

@type uuid4() :: bitstring()