Delta.Commit behaviour (Delta v0.1.0)

Internal commits API

An interface to work on top of Delta.DataLayer which should also implement Delta.Commit

Types

id()

@type id() :: Delta.uuid4() | t()

t()

@type t() :: %Delta.Commit{
  autosquash?: boolean() | nil,
  document_id: Delta.uuid4(),
  id: Delta.uuid4(),
  meta: any(),
  order: non_neg_integer() | nil,
  patch: Delta.Json.Patch.t(),
  previous_commit_id: Delta.uuid4() | nil,
  reverse_patch: Delta.Json.Patch.t() | nil,
  updated_at: DateTime.t() | nil
}

Represents a commit made by user to a document.

Fields

:id – UUIDv4 in default form. Required

:previous_commit_id – UUIDv4 in default form. Used to order commits of particular document to form a history. Required

:document_id – UUIDv4 in default from. Required

:order – Order of commit in document's history. Autogenerated

:autosquash? – Should the commits be squashed with the last commit with :patch with same paths

:patch – Changes to document in RFC 6902 json patch format. Required

:reverse_patch – RFC 6902 json patch to revert document to previous state. Autogenerated

:meta – any metadata, e. g. user who made it

:updated_at – when the commit was updated.

Autosquashing

Autosquashing is delta's feature for continuous and frequent document updates. It reduces the number of commits by squashing (see Delta.Commit.squash/2). If there will be a commit with autosquahs?: false and its delta will have any path which is being autosquahed at the moment, the commit will not be squashed. In order for two commits to be autosquashed, they must be marked autosquash?: true and have delta with same paths. For autosquash commit no checks of previous_commit_id are performed.

Callbacks

delete(layer_id, id, continuation?)

@callback delete(Delta.DataLayer.layer_id(), id(), continuation? :: boolean()) ::
  {:atomic, :ok, Delta.DataLayer.continuation()}

Same as Delta.Commit.delete/1, but returns data with continuation

get(layer_id, id, continuation?)

@callback get(Delta.DataLayer.layer_id(), id(), continuation? :: boolean()) ::
  {:atomic, t(), Delta.DataLayer.continuation()}
  | {:aborted, Delta.Errors.DoesNotExist.t(), Delta.DataLayer.continuation()}

Same as Delta.Commit.get/1, but returns data with continuation

list(layer_id, continuation?)

@callback list(Delta.DataLayer.layer_id(), continuation? :: boolean()) ::
  {:atomic, [t()], Delta.DataLayer.continuation()}

Same as Delta.Commit.list/1, but returns data with continuation

list(layer_id, arg2, arg3, continuation?)

@callback list(
  Delta.DataLayer.layer_id(),
  id() | nil,
  id() | nil,
  continuation? :: boolean()
) ::
  {:atomic, [t()], Delta.DataLayer.continuation()}

Same as Delta.Commit.list/2, but returns data with continuation

squash(layer_id, id, id, continuation?)

@callback squash(Delta.DataLayer.layer_id(), id(), id(), continuation? :: boolean()) ::
  {:atomic, t(), Delta.DataLayer.continuation()}
  | {:aborted, Delta.Errors.DoesNotExist.t(), Delta.DataLayer.continuation()}

Same as Delta.Commit.squash/2, but returns data with continuation

write(layer_id, t, continuation?)

@callback write(Delta.DataLayer.layer_id(), t(), continuation? :: boolean()) ::
  {:atomic, t(), Delta.DataLayer.continuation()}
  | {:aborted, Delta.Errors.DoesNotExist.t() | Delta.Errors.AlreadyExist.t(),
     Delta.DataLayer.continuation()}

Same as Delta.Commit.write/1, but returns data with continuation

write_many(layer_id, list, continuation?)

@callback write_many(Delta.DataLayer.layer_id(), [t()], continuation? :: boolean()) ::
  {:atomic, [t()], Delta.DataLayer.continuation()}
  | {:aborted, Delta.Errors.DoesNotExist.t() | Delta.Errors.AlreadyExist.t(),
     Delta.DataLayer.continuation()}

Same as Delta.Commit.write_many/1, but returns data with continuation

Functions

add_commits(commits)

@spec add_commits([t()]) ::
  {:atomic, [t()]} | {:aborted, Delta.Errors.Conflict.t()} | {:aborted, any()}

Adds commits to a document's history and checks for history consistency.

Checks for conflicts of the first commit.

If the commit has resolvable conflict (commits's delta does not overlaps with deltas of all conflicting commits), resolves the conflict and writes the commit. If the commit has no conflicts, writes the commit.

If the commit has unresolvable conflict, aborts with Delta.Errors.Conflict.t().

delete(change_id)

@spec delete(id()) :: {:atomic, :ok}

Deletes commit with id = commit_id.

Returns {:atomic, :ok} even if commit with id = commit_id does not exist.

do_squash(commit1, commit2)

@spec do_squash(t(), t()) :: t()

Squashing implementation. See squash/2.

get(document_id, commit_id)

@spec get(Delta.Document.id(), id()) ::
  {:atomic, t()} | {:aborted, Delta.Errors.DoesNotExist.t()}

id(id)

@spec id(id() | nil) :: Delta.uuid4() | any()

Gets id from id()

list(document_id)

@spec list(Delta.Document.id()) :: {:atomic, [t()]}

Lists commits of Delta.Document with id = document_id.

Expensive operation. If document does not exists, returns empty list

list(document_id, from_commit_id, to_commit_id)

@spec list(Delta.Document.id(), id(), id()) :: {:atomic, [t()]}

Lists commit from newest – from_commit_id to oldest – to_commit_id.

If commit with id = from_commit_id does not exist, assumes it to be the latest commit. If commit with id = to_commit_id does not exist, assumes it to be the first commit.

overlap?(commit1, commit2)

@spec overlap?(t(), t()) :: boolean()

Checks if two commits overlap

Commits are overlapping if their patches overlap

resolve_conflicts(c, history)

@spec resolve_conflicts(commits :: [t()], history :: [t()]) ::
  {:ok, [t()]} | {:error, Delta.Errors.Conflict.t()}

Checks if commits have conflict(s) with history and resolves them if possible.

Commits must be sorted by theirs (would be) ascending order – first commit is the first element of the list. History must be sorted by descending order – last commit is the first element of the list.

squash(commit_id_1, commit_id_2)

@spec squash(id(), id()) :: {:atomic, t()} | {:aborted, Delta.Errors.DoesNotExist.t()}

Squashes Delta.Commit with id = commit_id_2 into one with id = commit_id_1.

Resulting commit will have metadata of the second commit.

The second commit may not exist.

Aborts with %Delta.Errors.DoesNotExist{} if commit with id = commit_id_1 or id = commit_id_2 does not exist.

validate(c)

@spec validate(t() | any()) :: {:ok, t()} | {:error, Delta.Errors.Validation.t()}

Validates commit according to the following rules:

:id – must be UUIDv4 in default form

:previous_commit_id – must be UUIDv4 in default form of previous commit or nil

:patch – must be valid RFC 6092 Json delta

:document_id – must be valid UUIDv4 of document in default form.

Note: other functions expect valid input, therefor before passing data to them it should be validated.

validate_many(list)

@spec validate_many([t() | any()]) ::
  {:ok, [t()]} | {:error, Delta.Errors.Validation.t()}

Validates commits according to the following rules:

Commits must be sequential, i. e. commit[i].id = commit[i + 1].previous_id

All commits must have same :document_id

Each commit must be valid (See validate/1)

write(commit)

@spec write(t()) :: {:atomic, t()} | {:aborted, any()}

@spec write([t()]) :: {:atomic, [t()]} | {:aborted, any()}

Writes commit.

write_many(commits)

Writes commits.