Delta.Commit.CacheLayer (Delta v0.1.0)

Caching layer for Delta.Commit

Works on top of mnesia.

When started, creates mnesia table :"#{Delta.Commit.CacheLayer}.#{document_id}" on the node it was started on. All operations are performed on this table.

All operations are accumulated and then periodically replicated on persistent data layer.

Functions

add_commits(layer_id, commits, continuation?)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

crash_handler(map)

Returns anonymous function /0, which deletes mnesia table used by the layer.

delete(layer_id, id, continuation?)

Deletes commit. Always successful. Continuation deletes commit on another data layer.

See Delta.Commit.delete/1

get(layer_id, id, continuation?)

Gets commit.

If it exists, continuation is always nil

See Delta.Commit.get/1

list(layer_id, continuation?)

Lists all commits on this data layer.

Continuation lists data on another data layer with priority to this data layer.

See Delta.Commit.list/1

list(layer_id, from, to, continuation?)

Lists commits from one to another.

Continuation lists data on another data layer with priority to this data layer.

See Delta.Commit.list/3

start_link(document_id, _ \\ nil)

Starts this DataLayer with specific document id.

Has no options.

write(layer_id, commit, continuation?)

Writes commit.

Continuation writes commit on another data layer.

See Delta.Commit.write/1

write_many(layer_id, commits, continuation?)

Writes a list of commits. Commits are assumed to have equal document_id.

Continuation writes commits on another data layer

See Delta.Commit.write_many/1