Delta.DataLayer behaviour (Delta v0.1.0)

Behaviour of Delta Data Layers.

Data layer is a GenServer, managing operations with entites associated with document. Each instance is expected to:

control its uniuqe portion of data.

join Swarm group Delta.DataLayer for proper layer_id form resolution.

Types

continuation()

@type continuation() :: (layer_id() -> any()) | {atom(), [any()]} | nil

Specifies operation continuation on another data layer. Is executed on data layer with layer_id.

layer_id()

@type layer_id() :: {module(), Delta.uuid4()} | pid() | {module(), Delta.Document.t()}

Identifier for layer instance, can be one of:

{module(), Delta.uuid4()}, where module() is data layer module (normal form)

pid() of data layer process

{module(), Delta.Document.t()} – same as normal form, but id will be extracted from a document.

Note that any Delta.uuid4() is valid id in type context, in other words, there is no checks if document exists to be performed.

Callbacks

crash_handler(any)

@callback crash_handler(any()) :: (... -> any())

Returns function that should be called on process which monitors data layer.

start_link(document_id, opts)

@callback start_link(document_id :: Delta.Document.t() | Delta.uuid4(), opts :: keyword()) ::
  {:ok, pid()} | {:error, any()}

Starts data layer for isolating any data related to document.

Functions

continue(layer_id, continuation)

@spec continue(layer_id(), continuation()) :: any()

Runs continuation on specified layer

layer_id_normal(layer_id)

@spec layer_id_normal(layer_id()) :: layer_id()

Converts layer_id to normal form.

layer_id_pid(pid)

@spec layer_id_pid(layer_id()) :: pid()

Converts layer_id to pid form.