Delta.Document (Delta v0.1.0)

Internal documents API

Types

collection()

@type collection() :: String.t()

id()

@type id() :: Delta.uuid4() | t()

t()

@type t() :: %Delta.Document{
  collection: String.t(),
  data: any(),
  id: Delta.uuid4(),
  updated_at: DateTime.t()
}

Functions

add_changes(document_id, changes)

@spec add_changes(id(), [Delta.Change.t()]) ::
  {:atomic, [Delta.Change.t()]}
  | {:aborted, Delta.Errors.DoesNotExist.t() | Delta.Errors.Conflict.t()}

Adds list of changes changes to document with id = document_id in transactional manner.

Aborts with %Delta.Errors.Conflict{} if changes do not form linear history or one or more changes conflict with existing changes.

Aborts with %Delta.Errors.DoesNotExist{} if document with id = document_id does not exist.

create(document)

@spec create(t()) :: {:atomic, t()} | {:aborted, Delta.Errors.AlreadyExists.t()}

Creates document.

Aborts with %Delta.Errors.AlreadyExists{} if document with id = document.id already exist.

delete(document_id)

@spec delete(id()) :: {:atomic, :ok} | {:aborted, reason :: any()}

Deletes document and its changes with id = document_id. Reutrns {:atomic, :ok} even if document with id = document_id does not exist.

get(document_id)

@spec get(id()) :: {:atomic, t()} | {:aborted, Delta.Errors.DoesNotExist.t()}

list()

@spec list() :: {:atomic, [Delta.uuid4()]} | {:aborted, reason :: any()}

Returns ids of all documents for efficiency reasons

list(collection)

@spec list(collection()) :: {:atomic, [t()]} | {:aborted, reason :: any()}

Returns ids of all documents in collection for efficiency reasons

update(document_id, attrs \\ [])

@spec update(id(), map() | keyword()) ::
  {:atomic, t()} | {:aborted, Delta.Errors.DoesNotExist.t()}

Updates document.

Aborts with %Delta.Errors.DoesNotExist{} if document with id = document.id does not exist.

validate(document)

@spec validate(t()) :: {:atomic, t()} | {:aborted, Delta.Errors.Validation.t()}

Validates document.id to be UUIDv4 in default form