Delta.Json.Patch (Delta v0.1.0)

Helper functions for Json Patch (RFC 6902)

Note that strictly speaking it is not RFC 6902-compliant.

appliance of Delta.Json.Patch to data can never fail

test operation is ignored

replace and add considered to be the same operations

This is due to optimisation reasons.

Types

add()

@type add() :: {:add, Json.Pointer.t(), any()}

Adds a value and creates all that is required to make the pointer valid.

copy()

@type copy() :: {:copy, Json.Pointer.t(), from :: Json.Pointer.t()}

Copies the value from a path to another path. If the from path cant be resolved, the copied value will be nil.

move()

@type move() :: {:move, Json.Pointer.t(), from :: Json.Pointer.t()}

Moves the value from a path to another path. If the from path cant be resolved, the moved value will be nil.

operation()

@type operation() :: add() | remove() | move() | copy()

remove()

@type remove() :: {:remove, Json.Pointer.t()}

Removes value from path. NOOP if the value does not exists.

t()

@type t() :: [operation()]

List of operations that will be applied in order they are defined.

Functions

normalize(patch)

@spec normalize(t()) :: t()

If a paths of two operations are the same, keeps the last one

overlap?(patch1, patch2)

Returns true if two patches have operations on overlapping paths

parse(s)

@spec parse(String.t() | list()) :: {:ok, t()} | {:error, any()}

Parses json patch into Delta.Json.Patch.t()

Note:

test operation is ignored

replace and add considered to be the same operations

squash(patch1, patch2)

@spec squash(t(), t()) :: t()

Joins two patches together